Backend/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.2
django-rest-swagger==2.0.7
django-taggit==0.22.2
django-taggit-serializer==0.1.5
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
